+++ src/components/viewer/CombinedModel.tsx
