+++ src/components/viewer/Legend.tsx
