+++ src/components/viewer/WireframeModel.tsx
